backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pgvector
asyncpg
python-multipart
pyjwt
passlib[bcrypt]
bcrypt==4.1.3
# Alembic migrations
alembic
# AI embedding (pgvector)
sentence-transformers
torch --index-url https://download.pytorch.org/whl/cpu
# IP geolocation (PostGIS)
requests
# Test
pytest
httpx
pytest-cov
